Add a safety measure to prevent a borked system
Using birb with arguments like "." or "./" would cause birb to start
removing all fakeroot directories, basically nuking the system. This is
now prevented by making sure that the package name doesn't include these
said characters - and a few others that shouldn't be in the package
names in the first place.

To counteract situations like these, there's also some work being done
to add a --repair option to birb that would reinstall packages that have
their fakeroots missing, but are supposedly installed according to the
birb_db